stock-ticker-project/data-processor/requirements.txt
kafka-python-ng
psycopg2-binary
redis
